back/requirements.txt
cryptography==38.0.3
black==22.10.0
Flask==2.2.2
Flask-Cors==3.0.10
Pillow==9.3.0
requests==2.28.1
